Fix the 27 cube Trapped in problem.
At last version，when I rotate one layer of the MagicCube,the 27 cubes
in this layer will trap into the layer center a little.
The cause of this problem, when I do rotate ,I do translation first,and
then rotate it.But in this way ,I need to caculate the position first ,
this may have precision problem.
For this , I add a translation_after_rotate variable to SceneObject.
the render order is
translation-->rotation-->translation_after_rotation . translation used
to allocate the center of the SceneObject.